jmap: just use unsigned long, not size_t
Judy.h uses "Word_t" which it defines to an "unsigned long", so just use that.

If I were writing Judy from scratch, I'd use size_t or uintptr_t.

+++ ccan/jmap/test/run-access-count.c
@@ -8,7 +8,7 @@
 int main(int argc, char *argv[])
 {
 	struct jmap *map;
-	size_t *value;
+	unsigned long *value;
 	int status;
 
 	plan_tests(9);
